Moved NmsNBT
