+++ src/main/java/team/unstudio/udpl/nms/NmsHelper.java
@@ -4,6 +4,7 @@ import org.bukkit.inventory.ItemStack;
 
 import team.unstudio.udpl.core.nms.asm.AsmNmsManager;
 import team.unstudio.udpl.nms.inventory.NmsItemStack;
+import team.unstudio.udpl.nms.nbt.NmsNBT;
 
 public final class NmsHelper {
